# swagger/swaggerv1.yaml
swagger: '2.0'
info:
  description: "Programmers Internship Server Assignment"
  version: 0.0.1
  title: To Do List 
  contact:
    email: lenon461@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8016'
basePath: /
tags:
  - name: job
    description: test
    externalDocs:
      description: Find out more
      url: 'http://swagger.io'
schemes:
  - http
paths:
  /job:
    get:
      tags:
        - job
      summary: getAlljob
      description: ''
      operationId: test
      responses:
        '200':
          description: success
    post:
      tags:
        - job
      summary: postNewjob
      consumes:
        - application/json
      produces:
        - application/json
      description: ''
      parameters:
        - in: body
          name: body
          schema:
              $ref: '#/definitions/job'
      responses:
        '200':
         description: success
    
  /job/{id}:
    get:
      tags:
        - job
      summary: getOnejob
      description: ''
      operationId: test
      parameters:
        - in: path
          name: id
      responses:
        '200':
          description: success
    put:
      tags:
        - job
      summary: modify one job
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
          - in: path
            name: id
          - in: body
            name: body
            schema:
                $ref: '#/definitions/job'
      responses:
        '200':
         description: success
        '302':
         description: success
    delete:
        tags:
          - job
        summary: delete One Job
        parameters:
            - in: path
              name: id
              schema:
                  $ref: '#/definitions/job'
        responses:
          '200':
           description: success
  /api/jobs:
    get:
      tags:
        - job
      summary: getAlljob
      description: ''
      operationId: test
      responses:
        '200':
          description: success
  /api/job/{id}:
    get:
      tags:
        - job
      summary: getOnejob
      description: ''
      parameters:
        - in: path
          name: id
      responses:
          '200':
             description: success
    post:
      tags:
        - job
      summary: postNewjob
      consumes:
        - application/json
      produces:
        - application/json
      description: ''
      parameters:
        - in: body
          name: body
          schema:
              $ref: '#/definitions/job'
      responses:
        '200':
         description: success
    
  /job/{id}:
    get:
      tags:
        - job
      summary: getOnejob
      description: ''
      operationId: test
      parameters:
        - in: path
          name: id
      responses:
        '200':
          description: success
    put:
      tags:
        - job
      summary: modify one job
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
          - in: path
            name: id
          - in: body
            name: body
            schema:
                $ref: '#/definitions/job'
      responses:
        '200':
         description: success
        '302':
         description: success
    delete:
        tags:
          - job
        summary: delete One Job
        parameters:
            - in: path
              name: id
              schema:
                  $ref: '#/definitions/job'
        responses:
          '200':
           description: success

    
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  job:
      type: object
      properties:
          title:
              type: string
          contents:
              type: string
          #deadline:
          #    type: string
          #    format: date
          complete:
              type: boolean
          priority:
              type: number
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'